feat: rename rebuild effect widget file
