+++ packages/flutter_causality/lib/flutter_causality.dart
@@ -1,5 +1,5 @@
 /// flutter_causality
 library flutter_causality;
 
-export 'src/effect_widget.dart';
+export 'src/effects/rebuild_effect_widget.dart';
 export 'src/universe_widget.dart';
